requirements.txt
anthropic>=0.40.0
pyyaml>=6.0
rich>=13.0
httpx>=0.27.0
openai>=1.0.0
